feat (refs DPLAN-12571): Add new Route for readonly list of original statements and move readonly list of statements to new StatementListController. Adjust twig to render only statements or only original statements, rename procedure to procedureId
