+++ demosplan/DemosPlanCoreBundle/Controller/Statement/DemosPlanStatementController.php
@@ -2311,41 +2311,6 @@ class DemosPlanStatementController extends BaseController
     }
 
     /**
-     * List all statements per procedure
-     * without any possibilities to edit.
-     *
-     * @throws ProcedureNotFoundException
-     * @throws Exception
-     *
-     * @DplanPermissions("area_admin_statement_list")
-     */
-    #[Route(name: 'dplan_procedure_statement_list', methods: ['GET'], path: '/verfahren/{procedureId}/einwendungen', options: ['expose' => true])]
-    public function readOnlyStatementListAction(
-        string $procedureId,
-        ProcedureCoupleTokenFetcher $tokenFetcher,
-        ProcedureService $procedureService
-    ): Response {
-        $procedure = $procedureService->getProcedure($procedureId);
-
-        if (null === $procedure) {
-            throw ProcedureNotFoundException::createFromId($procedureId);
-        }
-
-        $isSourceAndCoupledProcedure = $tokenFetcher->isSourceAndCoupledProcedure($procedure);
-
-        return $this->renderTemplate(
-            '@DemosPlanCore/DemosPlanStatement/list_statements.html.twig',
-            [
-                'procedure'    => $procedureId,
-                'title'        => 'statements',
-                'templateVars' => [
-                    'isSourceAndCoupledProcedure' => $isSourceAndCoupledProcedure,
-                ],
-            ]
-        );
-    }
-
-    /**
      * Imports Statements from a xlsx-file.
      *
      * @throws ProcedureNotFoundException
